backend/services/order_srv/requirements.txt
peewee
pymysql
passlib
grpcio
grpcio-tools==1.43.0
protobuf
loguru
python-consul2
rocketmq-client-python
nacos-sdk-python
jaeger-client